PR for scheduler-aware-rm-policy addons project

+++ rm/rm-server/src/main/java/org/ow2/proactive/resourcemanager/nodesource/infrastructure/LocalInfrastructure.java
@@ -53,7 +53,7 @@ public class LocalInfrastructure extends InfrastructureManager {
 
     public static final int DEFAULT_NODE_NUMBER = Math.max(2, Runtime.getRuntime().availableProcessors() - 1);
 
-    public static final int DEFAULT_TIMEOUT = 30000;
+    public static final long DEFAULT_TIMEOUT = 30000;
 
     @Configurable(description = "Absolute path to credentials file\nused to add the node to the Resource Manager", credential = true)
     private Credentials credentials;
@@ -75,7 +75,7 @@ public class LocalInfrastructure extends InfrastructureManager {
     private AtomicInteger index;
 
     @Configurable(description = "in ms. After this timeout expired\nthe node is considered to be lost")
-    private int nodeTimeout = DEFAULT_TIMEOUT;
+    private long nodeTimeout = DEFAULT_TIMEOUT;
 
     @Configurable(description = "Additional ProActive properties")
     private String paProperties = "";
